Thread: TaskPool -> TaskManager with a much simpler/elegant design
TaskManager uses a blocking queue which save a lot of cpu resources when there is no task to run.
It might become an option later, the blocking version has a contention problem due to a not global mutex.
The concurrent queue might be ehanced with a random selection between N queue and by verifying that we choosed the most prioritary task.