RTTI: refactored *ALL* RTTI layer
FAtom is the replacement for old TMetaAtom<>, it's a simple untemplated class.
FAtom is responsible for manipulation of its stored pointer, using a type known only by TTypeTraits<>.
FAtom does nothing but wrapping the traits, which is stored using TInSituPtr<> for an efficient inplace storage of a pointer to an empty virtual class (aka only vfptr).
TTypeTraits<> is the only template in the code which is specialized for every type (except for boiler plate helpers).
This is a major change to previous implementation, which was massively templated and very heavy to instantiate.

Long compilation times was indeed the true motivation behind this major refactor.
But it will have an heavy cost : almost everything in Core.Serialize must be rewritten after this commit.
One fundamental difference which is drastically simplying the code is that FAtom don't own the memory, it just manipulates a pointer.
This make creating an atom very efficient, but also changes radically lifetime management.
But all theses insvestments are FTW hopefuly : new code is much more elegant, simpler and efficient.

Automatic atom type promotion isn't implemented yet.
Totally dropped RTTI::FOpaqueData, waiting for a concrete usage, same goes for TDynamicClass<> whose utility is still to be proven.