Grammar: fixed implicit promotion rules
previous was brute forcing every promotion and stopped a first success, turns out this strategy's success is highly dependent on the order you test promotion.

to avoid this issue the newer method uses RTTI constexpr traits to determine is a type is eight bool, signed, unsigned, floating point or string.