Allocator: upped allocation hash to 128 bits in leak detector
had a suspicion of collision, so 64 bits might not be safe enough since the allocation trie is not keeping a copy of callstack.
separated callstack fingeprints from callstack headers in 2 different allocations, since fingerprints are not needed for enumeration.
it did double the memory used to store different callstacks from 8mb to 16mb (max is stil 524288).