Serialize: FParseList uses a linear heap instead of a vector
the previous TVector<> grew upto 4mb of contiguous data ever reallocating (identified by new memory tracking).

so the class was refactored to use a FLinearHeap for parsed tokens, and also store the value strings in the heap.
the linear heap won't ever reallocate and will do a better job at amortizing allocations than a pool in this case (due to the strings).
using freshly added FParseMatch which is like Lexer::FMatch but using a FStringView instead of a FString.
FParseMatch has a TIntrusiveListNode<> which represents the chain of parsed tokens, effectively replacing the previous TVector<> (iterative access still possible thanks to FParseList::Seek()).
FParseMatch is also trivially destructible, so we can dellocate all memory in one call to ReleaseAll() when done.