Container: fixed SPARSEARRAY_INSITU()
can't use a TSegregatorAllocator<> + TInSituAllocator<> combo there, since it's relying on the assumption than the client only allocates 1 block
and TSparseArray<> will keep alive many blocks : 1 one each chunk (or less if reserved) + 1 for storing an array of chunks.

the solution was to use TFallbackAllocator<> with the new TInSituStackAllocator<> which tracks its state and can handle several allocations,
even if in this case it's just used for allocating 1 block with all the size available and then fallback on FMallocator.

One cons of this solution is the addtional 'Offset' from the stack allocator that we now need to store,
maybe it could be handled better if the allocator could gain info about the number of live allocations by passing it as argument (instead of storing another field).

#TODO ^^