HAL: lazily loading debug symbols from program counter
instead of loading *ALL* symbols from *ALL* modules loaded
could be slower, and could use a cache, but this is happening rarely atm (debug window).

also I was initializing the symbols inside CaptureCallstack(), which was wrong : it's not needed
when we capture the program counters from the stack, instead we want to load them when we decode.
this can make a big difference when we capture many callstack without decoding them.