Allocation: effective optimization of FMallocBinned thanks to VTune
While profiling the hash tables I found that one the main blottleneck for whole program was inside InitializeChunk_() from MallocBinned.cpp.
This function initialized the chunk when the size class did not match what we expected (when fresh or recycled), and needed to warm up the FreeBlocks.
It was the precise problem : that function would walk the whole chunk to setup the linked list, thus trashing all callee precious cache.
Originally I thought this function would be called only once when allocation a fresh chunk, but turns out that with chunk recycling it happens more often than I'd plan.

I fixed this issue by letting the FreeBlocks beeing lazily initialized : we store an addtional HighestIndex used when the FreeBlocks is null, which points to the highest block ever allocated in the chunk.
If this counter is != 0 we can deduce the address of a new block, without trashing CPU caches (since we don't really acces that  memory, we just craft a pointer address).
The FreeBlocks then only a block is deleted, which is fine, since at that moment the CPU might still be warm (provided a destructor was called on it).
Another round of VTune revealed that the bottleneck had disappeared completely, it's a quite big optimization than changed completely the bottlenecks of the program, which are now pointed at hash tables, as expected.
FMallocBinned is now even faster, and can't be seen in the benchmarks.

Thanks to VTune and micro architecture exploration, this tool is the real C++ profiler I've been expected, and now it's free :)