Build: stubbing new build system
it will be entirely described in ruby,
and bff will be generated.
this change will allow to switch back to vanilla fastbuild,
and stop maintaining my custom fork (mostly true).
it will also greatly simplify most of build tools I'm using,
but it needs a complete rewrite ^^;

this is still another journey