Task: FCompletionPort is now a FWeakRefCountable
Using a different counter from task counter,
it's important to dissociate task locking from lifetime locking to be thread-safe.
This had to be handled by each client previously,
but this is now much simpler thanks to TWeakPtr.
Could use a better allocation scheme in the future thanks to
custom deleter support inside TWeakPtr.