Thread: add new generalized FAtomicPhaseLock
based on recent work inside Test_Thread.cpp,
this atomic lock will trigger a transition hook when phase is updated
and make sure every thread are only released when the hook is finished.

it seems to be a concept worth of generalizing, as any dependency graph
can leverage this tool to enable incremental task subscription (and many more!).