Allocator: lower memory consumption for mipmap allocator
Previous implementation relied on wacky heuristics to determine when a new block
was needed, and only tracking the lest most block with largest memory block available.

I added a check to make sure that we were not wasting more than 60% of the allocated
memory, turns out most of the mipmaps were completely unused but the allocator was still
asking for more...

The algorithm needed a better tracking what was available, but I still wanted to keep
the main execution path atomic and didn't want to have to lock for a global queue.
I implemented the best I could thought of: using a static bitset with 1 bit for every
mip allocated, 1 meaning free memory available and 0 fully allocated.
Now when an allocation failed we look into this bitset for the leftmost bit set to 1,
and use the corresponding mip for the allocation. If the allocation failed (no more
space in the mipmap), then we simply recurse and ask for a new mip.