Allocator: mipmap allocator can garbage collect itself automatically
instead of manually, which was the previous approach.

This patch is saving (again) hundreds of mb during the unit test.
There's a performance penalty, but I chose to to lower the memory consumption vs
lowering memory contention. Still the current heuristic for automatic GC is
pretty forgiving and should let some free memory accumulate to ensure correct, and
thus maintain good performance.