Allocator: optimized Realloc/Relocate() for medium/large blocks
Mipmaps allows for shrinking a block in-place if the ptr is correctly, which has a near 100% hit rate in synthetic tests.

Mipmaps also allows for expanding a block in-place if the next blocks are free and if the ptr is correctly aligned, which has a good hit rate in synthetic tests.

Using both operations lowered memory usage (no more temporary block) and can avoid tons of embarrassing memcpy() calls on huge memory hunks.