Allocator: add TMipMapAllocator2<> as new default allocator for medium/large allocs
This is nearly the same algorithm than previous mipmap,
but not relying on a big contiguous chunk of reserved memory.
This added quite some complexity... but this is now a more generalist allocator.
I'm not I won't rollback, hence the 2 suffix.
I may also build upon this fork to change this algorithm to non-power of 2 alignment
since internal allocation is heavy atm (even if we mitigate this with alignment policies).