Allocator: first draft of new TBitmapHeap<>, which will replace TMipMapAllocator<>
TMipMapAllocator<> strategy to round to pow2 blocks is causing many internal fragmentation.
The new bitmap heap has much finer granularity.
It's working well but also that FMallocBinned is overbooking a lot of pages, and need a refactor
to lower memory consumption (and pressure on the bitmap heap).
FMallocBinned2 refactor is starting (rewrite of FMallocBinned).