Task: use ITaskContext instead of FTaskManager when possible
FTaskManager was used to handle non-fiber threads, but it made the API
far less flexible.

Every FTaskManager now expose a GlobalContext() method to get non-fiber
threads handling, and a new global ImmediateTaskContext() allows to easily
switch between threaded and non-thread dispatch.

This comes at the expanse of an additional VTable indirection, but compared
to the task management overhead it should be neglictable.