Memory: iterated on TMemoryPool/TCachedMemoryPool<> performance
TMemoryPool4<> mimicks FMallocBinned2 with an arbitrary number of
concurrent buckets and bundles smaller than a chunk: this prove to be the
only faster implementation compared to the original TMemoryPool<> (note
that the benchmark is really synthetical...)

On the other hand TCachedMemoryPool3<> is using a custom fixed size hash
table with closed addressing and a fine grained locking scheme: this
implementation is way faster than TCachedMemoryPool<> in multi-threaded
scenariis since the cache can now be updated from different threads at
the same time.

A future commit will remove other slower implementations and switch to the
new ones.