Memory: better polycies for TSlabHeap/TPoolingSlabHeap<>
slab size is not fixed, but will grow according to number of slabs already
allocated, or to the requested size if the slab is not large enough. this
way we don't have to handle spare blocks in the pooling variant.


the pooling variant have fewer buckets (max 4K for smaller slab sizes), but
can now reclaim memory from a wasted user alloc (see non-standalone blocks).
this recycling mechanism is way simpler than before.