[Allocator] track unaccounted memory in a specific memory domain
By now every allocation in the codebase should be accounted for by some
FMemoryTracking instance, but external code can still use the allocator
thanks to the overloaded malloc/free/etc operators.
The memory allocated can be tracked with a new method
SetThreadTrackingData(), and if the scope is not tagged the memory will
go to unaccounted. Previous codepath just need to call this function to
avoid this.
This information was added to investigate the memory usage reported in
visual studio for WindowTestApp.