Container: add new dense hashing tables TDenseHashSet<> and TDenseHashMap<> for fast iteration times with constant access time
Both are using the same fundamental algorithm: a sparse bucket list with a dense node list, using a prime number as sparse capacity to maximize hash entropy.
TDenseHashSet<> is the only one to support reverse iteration though (TDenseHashMap<> can't since it returns a pair of references instead of a reference to a pair).
At the moment they seem to be way slower than THashMap<> / THashSet<> counter-part for regular operations, which is expected to a certain measure since they are both optimized for dense iterations.