Malloc: implement a true tracking_malloc_aligned, since VK needs higher than 16
The previous implementation was relying on the natural allocation boundary set to 16,
but vulkan does need more than 16 (seen 64 and 256): I had to handle the case with a specialized header to store the original pointer.
Also fixed the disabled code path to revert to aligned_malloc/realloc/free and not the regular malloc() like this was previously.